Makeabilitylabwebsite banners (#49)
Added basic admin controls for uploading banner images, selecting which pages they will appear on, and adding title/caption text.

Updated the styles of the banner text overlays to be easier to read. Also attempted to add alt-text for background images through the use of the aria-label tag, but I’m not sure whether or not it will work as intended.