Added banner links and improved alt-text behavior
You can now specify a link for each banner on the admin side, which
will be visited when clicking on the image. You can still specify links
and other html tags inside the banner title/caption fields.

Also improved the behavior of the banner alt text, so that it will now
actually appear when you hover over the image.