Miscellaneous landing page publication improvements
Created a reusable citationPopover jQuery object, and updated both the
publications and landing pages to use it.

Capped the venue info to one line and the authors to three lines to
avoid overflow.

Fixed formatting of the gray border separating the download links.

Added best paper icons, etc. to the landing page.