Errant Commas addressed in People Listing
I believe that the issue found with the errant commas lies in fact that the active_phd passes both tests on lines 86 and 88 which causes two commas to be printed even though active_postdoc is false. I changed it so that the comma should only be printed if active_postdoc is true.