Working on addressing #1131
It looks like my prior code was not saving updated artifact fields correctly as I misunderstood how super.save() and update_fields were working.

if 'update_fields' is not provided (or if it is None) in 'kwargs', then all fields will be updated.

If it does exist, then we have to add in those fields that we modify ourselves in the save() function to 'update_fields', which I've now done, I believe.